requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.4
python-multipart==0.0.18
httpx==0.28.1

# Database
sqlalchemy==2.0.36
alembic==1.14.1

# AI / LLM
anthropic==0.42.0

# OCR
pytesseract==0.3.13
Pillow==11.1.0
PyPDF2==3.0.1
pdf2image==1.17.0

# File watching
watchdog==6.0.0

# Scheduling
APScheduler==3.10.4

# PDF generation
weasyprint==63.1

# Data processing
pandas==2.2.3

# CLI
click==8.1.8
rich==13.9.4

# Config
python-dotenv==1.0.1

# HTMX (served from static, no pip package needed)

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
